+++ tests/testthat/test-arima.R
@@ -169,13 +169,6 @@ test_that("REQUIRE TEST timeseries deprecation", {
     expect_warning(
         ts.out <- zelig(formula = unemp ~ leftseat, order = c(1, 0, 0), ts = "year",
                     data = subset, model = "arima"),
-        "All Zelig time series models will be deprecated on 1 February 2018"
+        "All Zelig time series models are deprecated"
     )
-
-    # Error to remind maintainers to deprecate
-    if (Sys.Date() > "2018-02-01")
-        expect_error(
-            ts.out <- zelig(formula = unemp ~ leftseat, order = c(1, 0, 0),
-                            ts = "year", data = subset, model = "arima")
-        )
 })